feat(moe): Promote routing precision to FP32/FP64 for fine-grained MoE (i.e., DeepSeek) training.

Closes #414 and #402

See merge request ADLR/megatron-lm!2677

+++ megatron/core/transformer/moe/fused_a2a.py
@@ -98,7 +98,7 @@ class FusedDispatch(torch.autograd.Function):
         ) = buffer.dispatch(
             x,
             topk_idx=token_indices,
-            topk_weights=token_probs.float(),
+            topk_weights=token_probs,  # DeepEP only supports float32 probs
             num_tokens_per_rank=num_tokens_per_rank,
             num_tokens_per_rdma_rank=num_tokens_per_rdma_rank,
             is_token_in_rank=is_token_in_rank,
